Chunk cache fully integrated

+++ include/akumuli.h
@@ -190,7 +190,7 @@ AKU_EXPORT apr_status_t aku_create_database
                                 // optional args
                                 , uint32_t  compression_threshold
                                 , uint64_t  window_size
-                                , uint32_t  max_cache_size
+                                , uint64_t  max_cache_size
                                 , aku_logger_cb_t logger
                                 );
